+++ CHANGELOG.md
@@ -4,6 +4,7 @@ The versions >= 1.21 are fork.
 
 ### Unreleased
 
+* Added regexp-based CSSParser
 * Fixed definition provider; it now works consistently with the completion provider
 * Fixed logging; it now writes to the Output panel (rather than developer tools) and shows error details
 * Upgraded dependencies